requirements.txt
streamlit==1.28.1
opencv-python==4.8.1.78
tensorflow==2.13.0
tensorflow-hub==0.14.0
Pillow==10.0.1
numpy==1.24.3
